+++ Practice/Baekjoon2884.py
