+++ Practice/numpy101.py
@@ -1,5 +1,4 @@
-import numpy as np
-from activation_functions.functions import *
+from Network.activation_functions.functions import *
 
 a = np.array([5, 4, 3, 2, 1])
 print(a[a % 2 == 0]) #boolean indexing
